+++ apps/io-wallet-user-func/src/app/config.ts
@@ -61,7 +61,7 @@ export type CryptoConfiguration = t.TypeOf<typeof CryptoConfiguration>;
 export const AttestationServiceConfiguration = t.type({
   allowedDeveloperUsers: t.array(t.string),
   androidBundleIdentifiers: t.array(t.string),
-  androidCrlUrl: t.string,
+  androidCrlUrls: t.array(t.string),
   androidPlayIntegrityUrl: t.string,
   androidPlayStoreCertificateHash: t.string,
   appleRootCertificate: t.string,
@@ -81,14 +81,19 @@ export type AttestationServiceConfiguration = t.TypeOf<
 const AzureConfig = t.type({
   appInsights: AzureAppInsightsConfig,
   cosmos: AzureCosmosConfig,
-  queue: t.type({
-    walletInstanceRevocation: t.type({
-      connectionString: t.string,
-      name: t.string,
-    }),
-  }),
   storage: t.type({
     entityConfiguration: t.type({ containerName: t.string }),
+    walletInstances: t.type({
+      connectionString: t.string,
+      queues: t.type({
+        pidIssuerRevokeApi: t.type({
+          name: t.string,
+        }),
+        validateCertificates: t.type({
+          name: t.string,
+        }),
+      }),
+    }),
   }),
 });
 
@@ -178,9 +183,10 @@ export const getAttestationServiceConfigFromEnvironment: RE.ReaderEither<
       RE.map((identifiers) => identifiers.split(",")),
       RE.orElse(() => RE.right(["it.pagopa.io.app"])),
     ),
-    androidCrlUrl: pipe(
-      readFromEnvironment("AndroidCrlUrl"),
-      RE.orElse(() => RE.right(ANDROID_CRL_URL)),
+    androidCrlUrls: pipe(
+      readFromEnvironment("AndroidCrlUrls"),
+      RE.map((identifiers) => identifiers.split(",")),
+      RE.orElse(() => RE.right([ANDROID_CRL_URL])),
     ),
     androidPlayIntegrityUrl: pipe(
       readFromEnvironment("AndroidPlayIntegrityUrl"),
@@ -235,9 +241,14 @@ export const getAzureConfigFromEnvironment: RE.ReaderEither<
     entityConfigurationStorageContainerName: readFromEnvironment(
       "EntityConfigurationStorageContainerName",
     ),
-    revocationQueueName: readFromEnvironment("RevocationQueueName"),
+    pidIssuerRevokeApiQueueName: readFromEnvironment(
+      "PidIssuerRevokeApiQueueName",
+    ),
     storageAccountConnectionString: readFromEnvironment(
       "StorageConnectionString",
+    ),
+    validateWalletInstanceCertificatesQueueName: readFromEnvironment(
+      "ValidateWalletInstanceCertificatesQueueName",
     ),
   }),
   RE.map(
@@ -245,20 +256,26 @@ export const getAzureConfigFromEnvironment: RE.ReaderEither<
       appInsights,
       cosmos,
       entityConfigurationStorageContainerName,
-      revocationQueueName,
+      pidIssuerRevokeApiQueueName,
       storageAccountConnectionString,
+      validateWalletInstanceCertificatesQueueName,
     }) => ({
       appInsights,
       cosmos,
-      queue: {
-        walletInstanceRevocation: {
-          connectionString: storageAccountConnectionString,
-          name: revocationQueueName,
-        },
-      },
       storage: {
         entityConfiguration: {
           containerName: entityConfigurationStorageContainerName,
+        },
+        walletInstances: {
+          connectionString: storageAccountConnectionString,
+          queues: {
+            pidIssuerRevokeApi: {
+              name: pidIssuerRevokeApiQueueName,
+            },
+            validateCertificates: {
+              name: validateWalletInstanceCertificatesQueueName,
+            },
+          },
         },
       },
     }),