+++ apps/io-wallet-user-func/src/app/main.ts
@@ -4,22 +4,25 @@ import withAppInsights from "@/infra/azure/appinsights/wrapper-handler";
 import { CosmosDbNonceRepository } from "@/infra/azure/cosmos/nonce";
 import { CosmosDbWalletInstanceRepository } from "@/infra/azure/cosmos/wallet-instance";
 import { AddWalletInstanceToValidationQueueFunction } from "@/infra/azure/functions/add-wallet-instance-to-validation-queue";
+import { CallPidIssuerRevokeApiFunction } from "@/infra/azure/functions/call-pid-issuer-revoke-api";
 import { CreateWalletAttestationFunction } from "@/infra/azure/functions/create-wallet-attestation";
 import { CreateWalletInstanceFunction } from "@/infra/azure/functions/create-wallet-instance";
 import { GenerateEntityConfigurationFunction } from "@/infra/azure/functions/generate-entity-configuration";
 import { GetCurrentWalletInstanceStatusFunction } from "@/infra/azure/functions/get-current-wallet-instance-status";
 import { GetNonceFunction } from "@/infra/azure/functions/get-nonce";
+import { GetWalletInstanceStatusFunction } from "@/infra/azure/functions/get-wallet-instance-status";
 import { HealthFunction } from "@/infra/azure/functions/health";
 import { SetCurrentWalletInstanceStatusFunction } from "@/infra/azure/functions/set-current-wallet-instance-status";
 import { SetWalletInstanceStatusFunction } from "@/infra/azure/functions/set-wallet-instance-status";
 import { ValidateWalletInstanceAttestedKeyFunction } from "@/infra/azure/functions/validate-wallet-instance-attested-key";
-import { WalletInstanceRevocationStorageQueue } from "@/infra/azure/queue/wallet-instance-revocation";
+import { WalletInstanceRevocationStorageQueue } from "@/infra/azure/storage/wallet-instance-revocation";
 import { CryptoSigner } from "@/infra/crypto/signer";
 import { PidIssuerClient } from "@/infra/pid-issuer/client";
 import { CosmosClient } from "@azure/cosmos";
 import { app, output } from "@azure/functions";
 import { DefaultAzureCredential } from "@azure/identity";
 import { QueueServiceClient } from "@azure/storage-queue";
+import { FiscalCode } from "@pagopa/ts-commons/lib/strings";
 import * as E from "fp-ts/Either";
 import { identity, pipe } from "fp-ts/function";
 import * as t from "io-ts";
@@ -53,13 +56,17 @@ const cosmosClient = new CosmosClient({
 });
 
 const queueServiceClient = QueueServiceClient.fromConnectionString(
-  config.azure.queue.walletInstanceRevocation.connectionString,
+  config.azure.storage.walletInstances.connectionString,
 );
 
 const revocationQueue = new WalletInstanceRevocationStorageQueue(
   queueServiceClient.getQueueClient(
-    config.azure.queue.walletInstanceRevocation.name,
+    config.azure.storage.walletInstances.queues.validateCertificates.name,
   ),
+);
+
+const pidIssuerRevokeApiQueue = queueServiceClient.getQueueClient(
+  config.azure.storage.walletInstances.queues.pidIssuerRevokeApi.name,
 );
 
 const database = cosmosClient.database(config.azure.cosmos.dbName);
@@ -151,6 +158,18 @@ app.timer("generateEntityConfiguration", {
   schedule: "0 0 */12 * * *", // the function returns a jwt that is valid for 24 hours, so the trigger is set every 12 hours
 });
 
+app.http("getWalletInstanceStatus", {
+  authLevel: "function",
+  handler: withAppInsights(
+    GetWalletInstanceStatusFunction({
+      telemetryClient: appInsightsClient,
+      walletInstanceRepository,
+    }),
+  ),
+  methods: ["GET"],
+  route: "wallet-instances/{id}/status",
+});
+
 app.http("getCurrentWalletInstanceStatus", {
   authLevel: "function",
   handler: withAppInsights(
@@ -159,7 +178,7 @@ app.http("getCurrentWalletInstanceStatus", {
       walletInstanceRepository,
     }),
   ),
-  methods: ["POST"],
+  methods: ["GET"],
   route: "wallet-instances/current/status",
 });
 
@@ -167,7 +186,7 @@ app.http("setWalletInstanceStatus", {
   authLevel: "function",
   handler: withAppInsights(
     SetWalletInstanceStatusFunction({
-      credentialRepository: pidIssuerClient,
+      queueClient: pidIssuerRevokeApiQueue,
       telemetryClient: appInsightsClient,
       walletInstanceRepository,
     }),
@@ -180,7 +199,7 @@ app.http("setCurrentWalletInstanceStatus", {
   authLevel: "function",
   handler: withAppInsights(
     SetCurrentWalletInstanceStatusFunction({
-      credentialRepository: pidIssuerClient,
+      queueClient: pidIssuerRevokeApiQueue,
       telemetryClient: appInsightsClient,
       walletInstanceRepository,
     }),
@@ -192,18 +211,18 @@ app.http("setCurrentWalletInstanceStatus", {
 app.cosmosDB("addWalletInstanceToValidationQueue", {
   connection: "CosmosDbEndpoint",
   containerName: "wallet-instances",
-  createLeaseContainerIfNotExists: true,
   databaseName: config.azure.cosmos.dbName,
   handler: AddWalletInstanceToValidationQueueFunction({
     inputDecoder: t.array(WalletInstance),
     telemetryClient: appInsightsClient,
   }),
-  leaseContainerName: "leases",
+  leaseContainerName: "leases-revoke-wallet-instance",
   leaseContainerPrefix: "wallet-instances-consumer-",
   maxItemsPerInvocation: 50,
   return: output.storageQueue({
     connection: "StorageConnectionString",
-    queueName: config.azure.queue.walletInstanceRevocation.name,
+    queueName:
+      config.azure.storage.walletInstances.queues.validateCertificates.name,
   }),
   startFromBeginning: true,
 });
@@ -218,5 +237,17 @@ app.storageQueue("validateWalletInstance", {
     telemetryClient: appInsightsClient,
     walletInstanceRepository,
   }),
-  queueName: config.azure.queue.walletInstanceRevocation.name,
+  queueName:
+    config.azure.storage.walletInstances.queues.validateCertificates.name,
+});
+
+app.storageQueue("callPidIssuerRevokeApi", {
+  connection: "StorageConnectionString",
+  handler: CallPidIssuerRevokeApiFunction({
+    credentialRepository: pidIssuerClient,
+    inputDecoder: FiscalCode,
+    telemetryClient: appInsightsClient,
+  }),
+  queueName:
+    config.azure.storage.walletInstances.queues.pidIssuerRevokeApi.name,
 });