+++ apps/io-wallet-user-func/src/infra/azure/functions/get-wallet-instance-status.ts
@@ -1,0 +1,6 @@
+import { GetWalletInstanceStatusHandler } from "@/infra/http/handlers/get-wallet-instance-status";
+import { httpAzureFunction } from "@pagopa/handler-kit-azure-func";
+
+export const GetWalletInstanceStatusFunction = httpAzureFunction(
+  GetWalletInstanceStatusHandler,
+);