+++ apps/io-wallet-user-func/src/wallet-instance.ts
@@ -6,6 +6,7 @@ import * as TE from "fp-ts/TaskEither";
 import { flow, pipe } from "fp-ts/function";
 import { EntityNotFoundError } from "io-wallet-common/error";
 import {
+  RevocationReason,
   WalletInstance,
   WalletInstanceValid,
   WalletInstanceValidWithAndroidCertificatesChain,
@@ -40,13 +41,28 @@ export interface WalletInstanceRepository {
   getValidByUserIdExcludingOne: (
     walletInstanceId: WalletInstance["id"],
     userId: WalletInstance["userId"],
-  ) => TE.TaskEither<Error, O.Option<WalletInstance[]>>;
+  ) => TE.TaskEither<Error, O.Option<WalletInstanceValid[]>>;
   insert: (walletInstance: WalletInstanceValid) => TE.TaskEither<Error, void>;
 }
 
 export interface WalletInstanceEnvironment {
   walletInstanceRepository: WalletInstanceRepository;
 }
+
+export const getWalletInstance: (
+  id: WalletInstance["id"],
+  userId: WalletInstance["userId"],
+) => RTE.ReaderTaskEither<WalletInstanceEnvironment, Error, WalletInstance> =
+  (id, userId) =>
+  ({ walletInstanceRepository }) =>
+    pipe(
+      walletInstanceRepository.get(id, userId),
+      TE.chain(
+        TE.fromOption(
+          () => new EntityNotFoundError("Wallet instance not found"),
+        ),
+      ),
+    );
 
 export const getCurrentWalletInstance: (
   userId: WalletInstance["userId"],
@@ -119,8 +135,9 @@ export const getValidWalletInstanceWithAndroidCertificatesChain: (
 export const revokeUserWalletInstances: (
   userId: WalletInstance["userId"],
   walletInstancesId: readonly WalletInstance["id"][],
+  reason: RevocationReason,
 ) => RTE.ReaderTaskEither<WalletInstanceEnvironment, Error, void> =
-  (userId, walletInstancesId) =>
+  (userId, walletInstancesId, reason) =>
   ({ walletInstanceRepository }) =>
     walletInstancesId.length
       ? walletInstanceRepository.batchPatch(
@@ -136,6 +153,11 @@ export const revokeUserWalletInstances: (
                 op: "add",
                 path: "/revokedAt",
                 value: new Date(),
+              },
+              {
+                op: "add",
+                path: "/revocationReason",
+                value: reason,
               },
             ],
           })),
@@ -169,14 +191,16 @@ const getUserValidWalletInstancesIdExceptOne: (
 export const revokeUserValidWalletInstancesExceptOne: (
   userId: WalletInstance["userId"],
   walletInstanceId: WalletInstance["id"],
+  reason: RevocationReason,
 ) => RTE.ReaderTaskEither<WalletInstanceEnvironment, Error, void> = (
   userId,
   walletInstanceId,
+  reason,
 ) =>
   pipe(
     getUserValidWalletInstancesIdExceptOne(userId, walletInstanceId),
     RTE.chain((validWalletInstances) =>
-      revokeUserWalletInstances(userId, validWalletInstances),
+      revokeUserWalletInstances(userId, validWalletInstances, reason),
     ),
   );