+++ infra/resources/_modules/apim/apim.tf
@@ -50,32 +50,9 @@ resource "azurerm_api_management_named_value" "user_func_key" {
   secret              = "true"
 }
 
-resource "azurerm_api_management_named_value" "trial_id" {
-  name                = "trial-id"
-  api_management_name = var.apim.name
-  resource_group_name = var.apim.resource_group_name
-  display_name        = "trial-id"
-  value               = "01J2GN4TA8FB6DPTAX3T3YD6M1"
-  secret              = false
-}
-
-resource "azurerm_api_management_named_value" "trial_system_api_key" {
-  name                = "trial-system-api-key"
-  api_management_name = var.apim.name
-  resource_group_name = var.apim.resource_group_name
-  display_name        = "trial-system-api-key"
-  value               = data.azurerm_key_vault_secret.trial_system_api_key.value
-  secret              = true
-}
-
 data "azurerm_key_vault_secret" "funciowallet_default" {
   name         = "funciowallet-KEY-APPBACKEND"
   key_vault_id = var.key_vault_id
-}
-
-data "azurerm_key_vault_secret" "trial_system_api_key" {
-  name         = "TrialSystemApiKey"
-  key_vault_id = var.key_vault_wallet_id
 }
 
 // SUPPORT API
@@ -98,9 +75,9 @@ module "apim_v2_wallet_support_api" {
 
   content_format = "openapi"
 
-  content_value = file("${path.module}/api/support-function/_swagger.json")
+  content_value = file("${path.module}/api/support/_swagger.json")
 
-  xml_content = file("${path.module}/api/support-function/_base_policy.xml")
+  xml_content = file("${path.module}/api/support/_base_policy.xml")
 }
 
 resource "azurerm_api_management_named_value" "support_func_key" {