+++ infra/resources/_modules/storage_accounts/outputs.tf
@@ -4,6 +4,18 @@ output "revocation_queue_name" {
   }
 }
 
+output "revocation_queue_name_02" {
+  value = {
+    name = azurerm_storage_queue.wallet_instances_revocation_check_02.name
+  }
+}
+
+output "pid_issuer_revoke_api_queue_name_01" {
+  value = {
+    name = azurerm_storage_queue.pid_issuer_revoke_api_01.name
+  }
+}
+
 output "wallet" {
   value = {
     id                            = module.storage_account.id