+++ apps/io-wallet-support-func/CHANGELOG.md
@@ -1,5 +1,45 @@
 # io-wallet-support-func
 
+## 1.4.2
+
+### Patch Changes
+
+- 9d44162: Added async call to PID issuer
+- Updated dependencies [9d44162]
+  - io-wallet-common@1.3.2
+
+## 1.4.1
+
+### Patch Changes
+
+- 1671ade: Updated turbo dependency version
+- Updated dependencies [1671ade]
+  - io-wallet-common@1.3.1
+
+## 1.4.0
+
+### Minor Changes
+
+- 94c97c3: Added migrateWalletInstances user function and updated repository dependencies
+
+### Patch Changes
+
+- Updated dependencies [94c97c3]
+  - io-wallet-common@1.3.0
+
+## 1.3.4
+
+### Patch Changes
+
+- d829f56: Add revocation_reason field to Wallet Instance
+
+## 1.3.3
+
+### Patch Changes
+
+- Updated dependencies [07d296c]
+  - io-wallet-common@1.2.1
+
 ## 1.3.2
 
 ### Patch Changes